[oh-tab-zhwi.9.2] Extract terminal session lifecycle module
